# Week4/hello/api/helloworld/v1/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Greeter API
    description: The greeting service definition.
    version: 0.0.1
paths:
    /v1/movie:
        get:
            tags:
                - Greeter
            operationId: Greeter_ListMovie
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListMovieReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        post:
            tags:
                - Greeter
            description: Sends a greeting
            operationId: Greeter_CreateMovie
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateMovieRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateMovieReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/movie/{id}:
        get:
            tags:
                - Greeter
            operationId: Greeter_GetMovie
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetMovieReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        put:
            tags:
                - Greeter
            operationId: Greeter_UpdateMovie
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int64
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateMovieRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateMovieReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        CreateMovieReply:
            type: object
            properties:
                Movie:
                    $ref: '#/components/schemas/Movie'
        CreateMovieRequest:
            type: object
            properties:
                title:
                    type: string
                geners:
                    type: string
        GetMovieReply:
            type: object
            properties:
                Movie:
                    $ref: '#/components/schemas/Movie'
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        ListMovieReply:
            type: object
            properties:
                result:
                    type: array
                    items:
                        $ref: '#/components/schemas/Movie'
        Movie:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                title:
                    type: string
                geners:
                    type: string
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        UpdateMovieReply:
            type: object
            properties:
                Movie:
                    $ref: '#/components/schemas/Movie'
        UpdateMovieRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                title:
                    type: string
                geners:
                    type: string
tags:
    - name: Greeter